refactor: Character 패키지
